PERF: avoid memory copies by not 'compacting' read buffers
In the recv thread data is read from the socket into (oversized) fix
sized buffers as it is available.  The buffers are then added to a
queue that a second thread pulls off of to interpret as CA and to
accumulate into the arrays that will eventually be passed to the
clients.  However, at each step the buffers are 'compacted', that is
if the last buffer in the list has space left, the contents of the new
buffer is copied into it.

This was presumably done to save memory, but it also means that we are
doing an extra, unneeded, copy of a vast majority of the data (the
fraction will be set by the ratio of the typical number of bytes
actually read from the socket and the buffer size).

This change simply removes the (pre-)compaction step.

We still have 1 copy (when we move from the fragmented buffers to the
single buffer in processIncoming)